suggestion(docs): don't specify a target server in the normative specification.
Since neither the Definition nor User Guide PDFs mention a required baseURL,
omitting the `servers` property may improve the clarity of the normative
OpenAPI document.

If OpenADR 3.0 _did_ have a desired pattern for a baseURL, the OpenAPI
specification could use a combination of `servers` and `serverVariables`
to specify the pattern. For more background, see\
https://swagger.io/specification/#server-object-example